+++ lib/features/home/presentation/widgets/home_tab_view.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class HomeTabView extends StatelessWidget {
+  final TabController tabController;
+  final List<Widget> tabViews;
+
+  const HomeTabView({
+    super.key,
+    required this.tabController,
+    required this.tabViews,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return TabBarView(
+      controller: tabController,
+      children: tabViews,
+    );
+  }
+}
